+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## WIP
 ### Added
 - `IRenderConfigC.VariableDeclarations` is now ignored if it only contains comments.
+- Simplified API for using `SmRunner`.
+  - https://github.com/StateSmith/StateSmith/issues/90
 
 ### Fixed
 - `StringUtils.RemoveCCodeComments()` now removes single line comments `//...` that end with string end (no line break).